move logic to content service
